+++ GlobalServerSoftware/src/main/java/com/elsys/globalserver/Services/CasualUsersService.java
@@ -6,6 +6,9 @@ import com.elsys.globalserver.DataAccess.CasualUserRepository;
 import com.elsys.globalserver.DB_Entities.CasualUser;
 import com.elsys.globalserver.DB_Entities.Prescription;
 import com.elsys.globalserver.DataAccess.PrescriptionsRepository;
+import com.elsys.globalserver.Services.Microservices.AuthenticationService;
+import com.elsys.globalserver.Services.Microservices.BugService;
+import com.elsys.globalserver.Services.Microservices.PrescriptionService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -14,50 +17,36 @@ import java.util.Set;
 
 @Service
 public class CasualUsersService {
-    private final CasualUserRepository casualUserRepository;
-    private final BugsRepository bugsRepository;
-    private final PrescriptionsRepository prescriptionsRepository;
+    private final AuthenticationService authenticationService;
+    private final PrescriptionService prescriptionService;
+    private final BugService bugService;
 
     @Autowired
-    public CasualUsersService(CasualUserRepository casualUserRepository,
-                              BugsRepository bugsRepository,
-                              PrescriptionsRepository prescriptionsRepository) {
-        this.casualUserRepository = casualUserRepository;
-        this.bugsRepository = bugsRepository;
-        this.prescriptionsRepository = prescriptionsRepository;
+    public CasualUsersService(AuthenticationService authenticationService,
+                              PrescriptionService prescriptionService,
+                              BugService bugService){
+        this.authenticationService = authenticationService;
+        this.prescriptionService = prescriptionService;
+        this.bugService = bugService;
     }
 
     public boolean register(CasualUser user_data) {
-        Optional<CasualUser> user = casualUserRepository.findByUsername(user_data.getUsername());
-
-        if (user.isPresent())
-            return false;
-
-        casualUserRepository.save(user_data);
-        return true;
+        return authenticationService.registerCasualUser(user_data);
     }
 
     public Optional<CasualUser> login(String username, String password) {
-        return casualUserRepository.findByUsernameAndPassword(username, password);
+        return authenticationService.loginCasualUser(username, password);
     }
 
     public Set<Prescription> getAllUserPrescriptions(int user_id) {
-        Optional<CasualUser> user = casualUserRepository.findById(user_id);
-        return user.map(CasualUser::getPrescriptions).orElse(null);
+        return prescriptionService.getUserPrescriptions(user_id);
     }
 
     public void reportBug(Bug bug) {
-        bugsRepository.save(bug);
+        bugService.reportBug(bug);
     }
 
     public boolean invalidatePrescription(int prescription_id){
-        Optional<Prescription> prescription = prescriptionsRepository.findById(prescription_id);
-
-        if (prescription.isEmpty())
-            return false;
-
-        prescription.get().setValid(false);
-        prescriptionsRepository.save(prescription.get());
-        return true;
+        return prescriptionService.invalidatePrescription(prescription_id);
     }
 }
